separate pixel functions from grid_detection.c and adapted tests
