+++ src/Preprocess/grid_detection.h
@@ -2,6 +2,7 @@
 #include <SDL2/SDL_image.h>
 #include "utils.h"
 #include "image_rotation.h"
+#include "pixel_utils.h"
 
 void grid_detection(SDL_Surface *image, int *x1, int *y1, int *x2, int *y2);
 
@@ -9,12 +10,6 @@ int *project(SDL_Surface *image, int axis);
 
 // aka écart type
 float standard_deviation(int *xs, int size);
-
-// returns a pointer to the pixel
-// at (x, y) in the image surface
-// or NULL if the pixel is out of bounds
-Uint32 *getpixel(SDL_Surface *surface, int x, int y);
-int is_pixel_white(SDL_Surface *image, int x, int y);
 
 // image: max convex component output
 // returns the angle at which the image is rotated
