+++ tests/PreprocessTests/test.c
@@ -1,2 +1,3 @@
 #include <criterion/criterion.h>
 #include "grid_detection_test.c"
+#include "pixel_utils_test.c"
